pyproject.toml
[project]
name = "winner_of_day"
version = "0.1.0"
description = "Default template for PDM package"
authors = [
    {name = "Daniil Shushakov", email = "4shushakov@gmail.com"},
]
dependencies = [
    "python-telegram-bot[job-queue]>=21.3",
    "click>=8.1.7",
    "requests>=2.32.3",
    "aiohttp>=3.9.5",
    "openai>=1.34.0",
]
requires-python = "==3.11.*"
readme = "README.md"
license = {text = "MIT"}


[build-system]
requires = ["pdm-backend"]
build-backend = "pdm.backend"

[tool.ruff]
line-length = 120
[tool.pdm]
distribution = true


